requirements.txt
# 代理系统依赖包

# Telegram Bot API
python-telegram-bot>=13.0,<14.0

# MongoDB驱动
pymongo>=3.11,<4.0

# 环境变量管理
python-dotenv>=0.19.0

# 数据处理
pandas>=1.3.0
xlsxwriter>=3.0.0

# HTTP请求
requests>=2.26.0

# 时间处理
pytz>=2021.1

# 日志
colorlog>=6.0.0
